feat: Add function to read addresses from Excel file
The code changes include adding a new function `read_addresses_from_excel` to the `df_improve.ipynb` notebook. This function reads addresses and coordinates from an Excel file using the `pandas` library. The purpose of this change is to provide a convenient way to import location data for display on a map.

Note: The recent user commits and repository commits are not relevant for generating the commit message in this case.